% Auto-generated: do not edit by hand
\name{dashIframeMessage}

\alias{dashIframeMessage}

\title{DashIframeMessage component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashIframeMessage(id=NULL, height=NULL, message=NULL, src=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{height}{Character. The height of the iframe}

\item{message}{Named list. The message from iframe.}

\item{src}{Character. The src of the iframe}

\item{width}{Character. The width of the iframe}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

